@@ -1,80 +1,40 @@
 {
-  "name": "team-finder-ui",
+  "name": "tf-replacement-ui",
+  "private": true,
   "version": "0.0.0",
-  "engines": {
-    "node": "^20",
-    "npm": "^10"
-  },
+  "type": "module",
   "scripts": {
     "dev": "vite",
-    "build": "vite build",
-    "serve": "vite preview",
-    "test": "jest",
-    "test:watch": "jest --watch",
-    "typecheck": "tsc",
-    "lint": "eslint src",
-    "lint:fix": "eslint src --fix",
-    "lint:report": "npm run lint -- --output-file eslint_report.json --format json src",
-    "storybook": "storybook dev -p 6006",
-    "build-storybook": "storybook build"
+    "build": "tsc && vite build",
+    "lint": "eslint . --ext ts,tsx --report-unused-disable-directives --max-warnings 0",
+    "preview": "vite preview"
   },
   "dependencies": {
-    "@sentry/react": "^7.50.0",
-    "@sentry/tracing": "^7.50.0",
-    "@types/react-modal": "^3.16.0",
-    "classnames": "^2.3.2",
-    "focus-visible": "^5.2.0",
+    "@tailwindcss/typography": "^0.5.10",
+    "formik": "^2.4.5",
+    "just-debounce-it": "^3.2.0",
     "react": "^18.2.0",
     "react-dom": "^18.2.0",
     "react-hot-toast": "^2.4.1",
-    "react-modal": "^3.16.1",
     "react-query": "^3.39.3",
-    "react-router-dom": "^6.11.0",
-    "react-select": "^5.7.2",
-    "react-svg": "^16.1.10",
-    "react-use": "^17.4.0"
+    "react-router-dom": "^6.21.1",
+    "react-select": "^5.8.0",
+    "react-svg": "^16.1.32"
   },
   "devDependencies": {
-    "@babel/core": "^7.21.5",
-    "@babel/preset-env": "^7.21.5",
-    "@babel/preset-react": "^7.18.6",
-    "@babel/preset-typescript": "^7.21.5",
-    "@storybook/addon-essentials": "^7.0.8",
-    "@storybook/addon-interactions": "^7.0.8",
-    "@storybook/addon-links": "^7.0.8",
-    "@storybook/blocks": "^7.0.8",
-    "@storybook/react": "^7.0.8",
-    "@storybook/react-vite": "^7.0.8",
-    "@storybook/testing-library": "^0.1.0",
-    "@tailwindcss/typography": "^0.5.9",
-    "@testing-library/react": "^14.0.0",
-    "@types/javascript-time-ago": "^2.0.3",
-    "@types/jest": "^29.5.1",
-    "@types/react": "^18.2.45",
-    "@types/react-dom": "^18.2.1",
-    "@types/react-router-dom": "^5.3.3",
-    "@typescript-eslint/eslint-plugin": "^5.59.1",
-    "@typescript-eslint/parser": "^5.59.1",
-    "@vitejs/plugin-react": "^4.0.0",
-    "autoprefixer": "^10.4.14",
-    "babel-jest": "^29.5.0",
-    "eslint": "^8.39.0",
-    "eslint-plugin-import": "^2.27.5",
-    "eslint-plugin-react": "^7.32.2",
+    "@types/react": "^18.2.43",
+    "@types/react-dom": "^18.2.17",
+    "@typescript-eslint/eslint-plugin": "^6.14.0",
+    "@typescript-eslint/parser": "^6.14.0",
+    "@vitejs/plugin-react-swc": "^3.5.0",
+    "autoprefixer": "^10.4.16",
+    "eslint": "^8.55.0",
     "eslint-plugin-react-hooks": "^4.6.0",
-    "eslint-plugin-storybook": "^0.6.12",
-    "jest": "^29.5.0",
-    "jest-environment-jsdom": "^29.5.0",
-    "jsdom": "^22.0.0",
-    "msw": "^1.2.1",
-    "postcss-focus-visible": "^8.0.2",
-    "postcss-import": "^15.1.0",
-    "prop-types": "^15.8.1",
-    "rollup-plugin-visualizer": "^5.9.0",
-    "storybook": "^7.0.8",
-    "tailwindcss": "^3.3.2",
-    "typescript": "^5.0.4",
-    "vite": "^4.3.3",
-    "whatwg-fetch": "^3.6.2"
+    "eslint-plugin-react-refresh": "^0.4.5",
+    "postcss": "^8.4.33",
+    "postcss-import": "^16.0.0",
+    "tailwindcss": "^3.4.1",
+    "typescript": "^5.2.2",
+    "vite": "^5.0.8"
   }
 }
