@@ -1,24 +1,10 @@
-import "focus-visible";
-import React from "react";
-import ReactDOM from "react-dom";
-import { ReactQueryDevtools } from "react-query/devtools";
-import { Context } from "./Context";
-import { importMetaEnv } from "./utils/importMeta";
-import { App } from "./App";
-import "./index.css";
+import React from 'react'
+import ReactDOM from 'react-dom/client'
+import {App} from './App.tsx'
+import './index.css'
 
-if (importMetaEnv().PROD) {
-  import("../src/utils/init-sentry");
-}
-
-ReactDOM.render(
+ReactDOM.createRoot(document.getElementById('root')!).render(
   <React.StrictMode>
-    <Context>
-      <React.Suspense fallback={null}>
-        <App />
-      </React.Suspense>
-      <ReactQueryDevtools />
-    </Context>
+    <App />
   </React.StrictMode>,
-  document.getElementById("root")
-);
+)
