@@ -1,25 +1,13 @@
-import { defineConfig } from "vite";
-import react from "@vitejs/plugin-react";
-import visualizer from "rollup-plugin-visualizer";
+import { defineConfig } from 'vite'
+import react from '@vitejs/plugin-react-swc'
 
 // https://vitejs.dev/config/
-export default defineConfig(({ mode }) => {
+export default defineConfig(() => {
   return {
     server: {
       host: "0.0.0.0",
-      port: 3000,
+      port: 3001,
     },
     plugins: [react()],
-    build: {
-      sourcemap: mode === "production",
-      rollupOptions: {
-        plugins: [
-          mode === "production" &&
-            visualizer({
-              filename: "build-stats.html",
-            }),
-        ].filter((x) => x),
-      },
-    },
-  };
+  }
 });
